+++ scripts/embedding/eclipse/eclipse_logtalk_eco.sh
@@ -6,7 +6,7 @@
 ##   compiler and runtime and optionally an application.eco file with
 ##   a Logtalk application
 ## 
-##   Last updated on May 16, 2018
+##   Last updated on December 11, 2018
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -93,12 +93,12 @@ else
 fi
 
 # default values
-directory="$HOME/collect"
+directory="$(pwd -P)"
 paths="$LOGTALKHOME/paths/paths_core.pl"
 compile="false"
 
 print_version() {
-	echo "$(basename "$0") 0.10"
+	echo "$(basename "$0") 0.11"
 	exit 0
 }
 
@@ -116,7 +116,7 @@ usage_help()
 	echo
 	echo "Optional arguments:"
 	echo "  -c compile library alias paths in paths and settings files"
-	echo "  -d directory for intermediate and final results (default is $directory)"
+	echo "  -d directory for the generated Prolog files (default is the current directory)"
 	echo "  -p library paths file (default is $paths)"
 	echo "  -s settings file"
 	echo "  -l loader file for the application"
@@ -175,7 +175,14 @@ else
 fi
 
 mkdir -p "$directory"
-cd "$directory" || exit 1
+
+temporary=$(mktemp -d)
+if [[ ! "$temporary" || ! -d "$temporary" ]]; then
+  echo "Could not create temporary directory!"
+  exit 1
+fi
+
+cd "$temporary" || exit 1
 
 operating_system=$(uname -s)
 
@@ -207,19 +214,19 @@ echo ":- discontiguous('\$lgt_extends_protocol_'/3)." >> logtalk.pl
 echo ":- discontiguous('\$lgt_loaded_file_'/7)." >> logtalk.pl
 echo ":- discontiguous('\$lgt_included_file_'/4)." >> logtalk.pl
 
-eclipselgt$extension -e "logtalk_compile([core(expanding),core(monitoring),core(forwarding),core(user),core(logtalk),core(core_messages)],[optimize(on),scratch_directory('$directory')]),halt"
+eclipselgt$extension -e "logtalk_compile([core(expanding),core(monitoring),core(forwarding),core(user),core(logtalk),core(core_messages)],[optimize(on),scratch_directory('$temporary')]),halt"
 
 if [ "$compile" != "false" ] ; then
-	eclipselgt$extension -e "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$paths',[hook(expand_library_alias_paths),scratch_directory('$directory')]),halt"
+	eclipselgt$extension -e "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$paths',[hook(expand_library_alias_paths),scratch_directory('$temporary')]),halt"
 else
-	cp "$paths" "$directory/paths_lgt.pl"
+	cp "$paths" "$temporary/paths_lgt.pl"
 fi
 
 if [ "$settings" != "" ] ; then
 	if [ "$compile" != "false" ] ; then
-		eclipselgt$extension  -L iso -t user -e "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$settings',[hook(expand_library_alias_paths),optimize(on),scratch_directory('$directory')]),halt"
+		eclipselgt$extension  -L iso -t user -e "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$settings',[hook(expand_library_alias_paths),optimize(on),scratch_directory('$temporary')]),halt"
 	else
-		eclipselgt$extension -L iso -t user -e "logtalk_compile('$settings',[optimize(on),scratch_directory('$directory')]),halt"
+		eclipselgt$extension -L iso -t user -e "logtalk_compile('$settings',[optimize(on),scratch_directory('$temporary')]),halt"
 	fi
 	cat \
 		eclipse.pl \
@@ -248,14 +255,19 @@ else
 fi
 
 eclipse -L iso -t user -e "compile(logtalk,[debug:off,opt_level:1,output:eco]),halt"
-rm ./*.pl
+mv logtalk.eco "$directory"
 
 if [ "$loader" != "" ] ; then
-	mkdir -p "$directory/application"
-	cd "$directory/application" || exit 1
-	eclipse -L iso -t user -f "$directory/logtalk.eco" -e "set_logtalk_flag(clean,off),set_logtalk_flag(scratch_directory,'$directory/application'),logtalk_load('$loader'),halt"
-	cat $(ls -rt "$directory/application"/*.pl) > application.pl
+	mkdir -p "$temporary/application"
+	cd "$temporary/application" || exit 1
+	eclipse -L iso -t user -f "$directory/logtalk.eco" -e "set_logtalk_flag(clean,off),set_logtalk_flag(scratch_directory,'$temporary/application'),logtalk_load('$loader'),halt"
+	cat $(ls -rt "$temporary/application"/*.pl) > application.pl
 	eclipse -L iso -t user -f "$directory/logtalk.eco" -e "compile(application,[debug:off,opt_level:1,output:eco]),halt"
-	mv application.eco ..
-	rm ./*.pl
+	mv application.eco "$directory"
 fi
+
+function cleanup {
+	rm -rf "$temporary"
+}
+# register cleanup function to be called on EXIT signal
+trap cleanup EXIT