+++ scripts/embedding/sicstus/sicstus_logtalk_po.sh
@@ -6,7 +6,7 @@
 ##   compiler and runtime and optionally an application.po file with a
 ##   Logtalk application
 ## 
-##   Last updated on May 16, 2018
+##   Last updated on December 11, 2018
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -93,12 +93,12 @@ else
 fi
 
 # default values
-directory="$HOME/collect"
+directory="$(pwd -P)"
 paths="$LOGTALKHOME/paths/paths_core.pl"
 compile="false"
 
 print_version() {
-	echo "$(basename "$0") 0.10"
+	echo "$(basename "$0") 0.11"
 	exit 0
 }
 
@@ -116,7 +116,7 @@ usage_help()
 	echo
 	echo "Optional arguments:"
 	echo "  -c compile library alias paths in paths and settings files"
-	echo "  -d directory for intermediate and final results (default is $directory)"
+	echo "  -d directory for the generated .po files (default is the current directory)"
 	echo "  -p library paths file (default is $paths)"
 	echo "  -s settings file"
 	echo "  -l loader file for the application"
@@ -175,7 +175,14 @@ else
 fi
 
 mkdir -p "$directory"
-cd "$directory" || exit 1
+
+temporary=$(mktemp -d)
+if [[ ! "$temporary" || ! -d "$temporary" ]]; then
+  echo "Could not create temporary directory!"
+  exit 1
+fi
+
+cd "$temporary" || exit 1
 
 operating_system=$(uname -s)
 
@@ -192,19 +199,19 @@ fi
 cp "$LOGTALKHOME/adapters/sicstus.pl" .
 cp "$LOGTALKHOME/core/core.pl" .
 
-sicstuslgt$extension --goal "logtalk_compile([core(expanding),core(monitoring),core(forwarding),core(user),core(logtalk),core(core_messages)],[optimize(on),scratch_directory('$directory')]),halt."
+sicstuslgt$extension --goal "logtalk_compile([core(expanding),core(monitoring),core(forwarding),core(user),core(logtalk),core(core_messages)],[optimize(on),scratch_directory('$temporary')]),halt."
 
 if [ "$compile" != "false" ] ; then
-	sicstuslgt$extension --goal "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$paths',[hook(expand_library_alias_paths),scratch_directory('$directory')]),halt."
+	sicstuslgt$extension --goal "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$paths',[hook(expand_library_alias_paths),scratch_directory('$temporary')]),halt."
 else
-	cp "$paths" "$directory/paths_lgt.pl"
+	cp "$paths" "$temporary/paths_lgt.pl"
 fi
 
 if [ "$settings" != "" ] ; then
 	if [ "$compile" != "false" ] ; then
-		sicstuslgt$extension --goal "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$settings',[hook(expand_library_alias_paths),optimize(on),scratch_directory('$directory')]),halt."
+		sicstuslgt$extension --goal "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$settings',[hook(expand_library_alias_paths),optimize(on),scratch_directory('$temporary')]),halt."
 	else
-		sicstuslgt$extension --goal "logtalk_compile('$settings',[optimize(on),scratch_directory('$directory')]),halt."
+		sicstuslgt$extension --goal "logtalk_compile('$settings',[optimize(on),scratch_directory('$temporary')]),halt."
 	fi
 	cat \
 		sicstus.pl \
@@ -233,14 +240,19 @@ else
 fi
 
 sicstus$extension --goal "set_prolog_flag(discontiguous_warnings,off),compile(logtalk),save_files(logtalk,logtalk),halt."
-rm ./*.pl
+mv logtalk.po "$directory"
 
 if [ "$loader" != "" ] ; then
-	mkdir -p "$directory/application"
-	cd "$directory/application" || exit 1
-	sicstus$extension --goal "load_files('$directory/logtalk.po'),set_logtalk_flag(clean,off),set_logtalk_flag(scratch_directory,'$directory/application'),logtalk_load('$loader'),halt."
-	cat $(ls -rt "$directory/application"/*.pl) > application.pl
+	mkdir -p "$temporary/application"
+	cd "$temporary/application" || exit 1
+	sicstus$extension --goal "load_files('$directory/logtalk.po'),set_logtalk_flag(clean,off),set_logtalk_flag(scratch_directory,'$temporary/application'),logtalk_load('$loader'),halt."
+	cat $(ls -rt "$temporary/application"/*.pl) > application.pl
 	sicstus$extension --goal "load_files('$directory/logtalk.po'),set_prolog_flag(discontiguous_warnings,off),compile(application),save_files(application,application),halt."
-	mv application.po ..
-	rm ./*.pl
+	mv application.po "$directory"
 fi
+
+function cleanup {
+	rm -rf "$temporary"
+}
+# register cleanup function to be called on EXIT signal
+trap cleanup EXIT