+++ scripts/embedding/xsb/xsb_logtalk_xwam.sh
@@ -5,7 +5,7 @@
 ##   This script creates a XSB logtalk.xwam file
 ##   with the Logtalk compiler and runtime
 ## 
-##   Last updated on May 16, 2018
+##   Last updated on December 11, 2018
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -121,19 +121,21 @@ else
 fi
 
 # default values
-directory="$HOME/collect"
+directory="$(pwd -P)"
 paths="$LOGTALKHOME/paths/paths_core.pl"
+compile="false"
 
 print_version() {
-	echo "$(basename "$0") 0.10"
+	echo "$(basename "$0") 0.11"
 	exit 0
 }
 
 usage_help()
 {
 	echo 
-	echo "This script creates a XSB logtalk.xwam file with the Logtalk compiler"
-	echo "and runtime"
+	echo "This script creates a XSB logtalk.xwam file with the Logtalk compiler and"
+	echo "runtime and an optional application.pl file from an application source"
+	echo "code given its loader file."
 	echo
 	echo "Usage:"
 	echo "  $(basename "$0") [-c] [-d directory] [-p paths] [-s settings] [-l loader]"
@@ -142,7 +144,7 @@ usage_help()
 	echo
 	echo "Optional arguments:"
 	echo "  -c compile library alias paths in paths and settings files"
-	echo "  -d directory for intermediate and final results (default is $directory)"
+	echo "  -d directory for the generated .xwam files (default is the current directory)"
 	echo "  -p library paths file (default is $paths)"
 	echo "  -s settings file"
 	echo "  -l loader file for the application"
@@ -201,7 +203,14 @@ else
 fi
 
 mkdir -p "$directory"
-cd "$directory" || exit 1
+
+temporary=$(mktemp -d)
+if [[ ! "$temporary" || ! -d "$temporary" ]]; then
+  echo "Could not create temporary directory!"
+  exit 1
+fi
+
+cd "$temporary" || exit 1
 
 operating_system=$(uname -s)
 
@@ -218,19 +227,19 @@ fi
 cp "$LOGTALKHOME/adapters/xsb.pl" .
 cp "$LOGTALKHOME/core/core.pl" .
 
-xsblgt$extension -e "logtalk_compile([core(expanding),core(monitoring),core(forwarding),core(user),core(logtalk),core(core_messages)],[optimize(on),scratch_directory('$directory')]),halt."
+xsblgt$extension -e "logtalk_compile([core(expanding),core(monitoring),core(forwarding),core(user),core(logtalk),core(core_messages)],[optimize(on),scratch_directory('$temporary')]),halt."
 
 if [ "$compile" != "false" ] ; then
-	xsblgt$extension -e "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$paths',[hook(expand_library_alias_paths),scratch_directory('$directory')]),halt."
+	xsblgt$extension -e "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$paths',[hook(expand_library_alias_paths),scratch_directory('$temporary')]),halt."
 else
-	cp "$paths" "$directory/paths_lgt.P"
+	cp "$paths" "$temporary/paths_lgt.P"
 fi
 
 if [ "$settings" != "" ] ; then
 	if [ "$compile" != "false" ] ; then
-		xsblgt$extension -e "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$settings',[hook(expand_library_alias_paths),optimize(on),scratch_directory('$directory')]),halt."
+		xsblgt$extension -e "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$settings',[hook(expand_library_alias_paths),optimize(on),scratch_directory('$temporary')]),halt."
 	else
-		xsblgt$extension -e "logtalk_compile('$settings',[optimize(on),scratch_directory('$directory')]),halt."
+		xsblgt$extension -e "logtalk_compile('$settings',[optimize(on),scratch_directory('$temporary')]),halt."
 	fi
 	cat \
 		xsb.pl \
@@ -260,18 +269,19 @@ fi
 
 xsb -e "compile(logtalk),halt."
 
-rm ./*.pl
-rm ./*.P
-rm ./*_pl.xwam
-rm ./*_lgt.xwam
+mv logtalk.xwam "$directory"
 
 if [ "$loader" != "" ] ; then
-	mkdir -p "$directory/application"
-	cd "$directory/application" || exit 1
-	xsblgt$extension -e "set_logtalk_flag(clean,off),set_logtalk_flag(scratch_directory,'$directory/application'),logtalk_load('$loader'),halt."
-	cat $(ls -rt "$directory/application"/*.P) > application.P
+	mkdir -p "$temporary/application"
+	cd "$temporary/application" || exit 1
+	xsblgt$extension -e "set_logtalk_flag(clean,off),set_logtalk_flag(scratch_directory,'$temporary/application'),logtalk_load('$loader'),halt."
+	cat $(ls -rt "$temporary/application"/*.P) > application.P
 	xsblgt$extension -e "compile(application),halt."
-	mv application.xwam ..
-	rm ./*.P
-	rm ./*.xwam
+	mv application.xwam "$directory"
 fi
+
+function cleanup {
+	rm -rf "$temporary"
+}
+# register cleanup function to be called on EXIT signal
+trap cleanup EXIT