+++ scripts/embedding/yap/yap_logtalk_pl.sh
@@ -6,7 +6,7 @@
 ##   and runtime and optionally an application.pl file with a Logtalk
 ##   application
 ## 
-##   Last updated on May 16, 2018
+##   Last updated on December 11, 2018
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -93,13 +93,13 @@ else
 fi
 
 # default values
-directory="$HOME/collect"
+directory="$(pwd -P)"
 paths="$LOGTALKHOME/paths/paths_core.pl"
 hooks="$LOGTALKHOME/adapters/yaphooks.pl"
 compile="false"
 
 print_version() {
-	echo "$(basename "$0") 0.10"
+	echo "$(basename "$0") 0.11"
 	exit 0
 }
 
@@ -107,7 +107,7 @@ usage_help()
 {
 	echo 
 	echo "This script creates a YAP logtalk.pl file with the Logtalk compiler and"
-	echo "runtime and an optional application.qlf file from an application source"
+	echo "runtime and an optional application.pl file from an application source"
 	echo "code given its loader file."
 	echo
 	echo "Usage:"
@@ -117,7 +117,7 @@ usage_help()
 	echo
 	echo "Optional arguments:"
 	echo "  -c compile library alias paths in paths and settings files"
-	echo "  -d directory for intermediate and final results (default is $directory)"
+	echo "  -d directory for the generated Prolog files (default is the current directory)"
 	echo "  -p library paths file (default is $paths)"
 	echo "  -k hooks file (default is $hooks)"
 	echo "  -s settings file"
@@ -187,7 +187,14 @@ else
 fi
 
 mkdir -p "$directory"
-cd "$directory" || exit 1
+
+temporary=$(mktemp -d)
+if [[ ! "$temporary" || ! -d "$temporary" ]]; then
+  echo "Could not create temporary directory!"
+  exit 1
+fi
+
+cd "$temporary" || exit 1
 
 operating_system=$(uname -s)
 
@@ -204,19 +211,19 @@ fi
 cp "$LOGTALKHOME/adapters/yap.pl" .
 cp "$LOGTALKHOME/core/core.pl" .
 
-yaplgt$extension -g "logtalk_compile([core(expanding),core(monitoring),core(forwarding),core(user),core(logtalk),core(core_messages)],[optimize(on),scratch_directory('$directory')]),halt"
+yaplgt$extension -g "logtalk_compile([core(expanding),core(monitoring),core(forwarding),core(user),core(logtalk),core(core_messages)],[optimize(on),scratch_directory('$temporary')]),halt"
 
 if [ "$compile" != "false" ] ; then
-	yaplgt$extension -g "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$paths',[hook(expand_library_alias_paths),scratch_directory('$directory')]),halt"
+	yaplgt$extension -g "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$paths',[hook(expand_library_alias_paths),scratch_directory('$temporary')]),halt"
 else
-	cp "$paths" "$directory/paths_lgt.yap"
+	cp "$paths" "$temporary/paths_lgt.yap"
 fi
 
 if [ "$settings" != "" ] ; then
 	if [ "$compile" != "false" ] ; then
-		yaplgt$extension -g "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$settings',[hook(expand_library_alias_paths),optimize(on),scratch_directory('$directory')]),halt"
+		yaplgt$extension -g "logtalk_load(library(expand_library_alias_paths_loader)),logtalk_compile('$settings',[hook(expand_library_alias_paths),optimize(on),scratch_directory('$temporary')]),halt"
 	else
-		yaplgt$extension -g "logtalk_compile('$settings',[optimize(on),scratch_directory('$directory')]),halt"
+		yaplgt$extension -g "logtalk_compile('$settings',[optimize(on),scratch_directory('$temporary')]),halt"
 	fi
 	cat \
 		yap.pl \
@@ -246,15 +253,17 @@ else
 		> logtalk.pl
 fi
 
-rm ./*.yap
-rm yap.pl
-rm core.pl
-rm -f yaphooks.pl
+mv logtalk.pl "$directory"
 
 if [ "$loader" != "" ] ; then
-	mkdir -p "$directory/application"
-	cd "$directory/application" || exit 1
-	yap -g "consult('../logtalk'),set_logtalk_flag(clean,off),set_logtalk_flag(scratch_directory,'$directory/application'),logtalk_load('$loader'),halt"
-	cat $(ls -rt "$directory/application"/*.yap) > ../application.pl
-	rm ./*.yap
+	mkdir -p "$temporary/application"
+	cd "$temporary/application" || exit 1
+	yap -g "consult('$directory/logtalk'),set_logtalk_flag(clean,off),set_logtalk_flag(scratch_directory,'$temporary/application'),logtalk_load('$loader'),halt"
+	cat $(ls -rt "$temporary/application"/*.yap) > "$directory"/application.pl
 fi
+
+function cleanup {
+	rm -rf "$temporary"
+}
+# register cleanup function to be called on EXIT signal
+trap cleanup EXIT